Automated code formatting pass
All changes in this PR were automatically generated by a code formatting
script that is in the master branch.
The style is configurable so if there's something you don't like we can
discuss, but today there is not a consistent style across the repo, and
often not even consistent within a file.

Signed-off-by: Dave Thaler <dthaler@ntdev.microsoft.com>