Add doxygen documentation for libbpf apis
Eventually this should be upstreamed, but as a workaround the docs
are put into our repo in the meantime.

As discussed at
https://stackoverflow.com/questions/23798053/how-to-document-errno-value-with-doxygen
doxygen does not have a built-in way to document errno values set.  It is
possible by creating doxygen custom commands, but for now we use
the "exception" section as recommended at the bottom of that page.

Fixes #490

Signed-off-by: Dave Thaler <dthaler@microsoft.com>