Skip CI/CD build and tests for doc-only changes
Note that if a PR is updated via a merge commit, e.g., using
the "Update with merge commit" variation of the "Update branch" button
in github, instead of a rebase then the build/tests will be done.

As such, for doc-only changes, to use the "Update branch" button in
github, be sure to use the down arrow to change it to "Update with
rebase".

Fixes #1329

Signed-off-by: Dave Thaler <dthaler@microsoft.com>