Update scorecard workflow to latest (#2383)
Note that we tried this before in PR #1161 and had to revert it in PR #1174
but the issue it was blocked on seems to be fixed and this should fix
the various scorecard failures we are now seeing.

MsQuic which originally did the same as we did already snapped to latest
as can be seen at
https://github.com/microsoft/msquic/blob/main/.github/workflows/scorecards-analysis.yml
and its scorecard runs are passing.

Signed-off-by: Dave Thaler <dthaler@microsoft.com>