Backwards compatible fixes to bpf() syscall emulation (#3870)
* bpf(): do not return errors via errno

The Linux ABI returns all syscall errors via the function return value,
not via errno.

Fixes https://github.com/microsoft/ebpf-for-windows/issues/3749

* Allow detecting if bpf() command is not implemented

Use SetLastError to indicate to callers that a bpf() command is not
implemented. This avoids polluting the bpf() return value with
platform specific error returns while still allowing detection of
this important case.

* Add forwards and backwards compatibility to bpf() emulation

On Linux, bpf() accepts a bpf_attr which is larger than what the
syscall expects, as long as the unknown fields are all 0. It also
accepts a bpf_attr which is smaller than what it expects, by assuming
that all missing fields are zero.

This allows forwards and backwards compatilibity between old and new
versions of both the Linux kernel and user space tooling.

Implement a similar scheme for the bpf() emulation.

* Return EPERM from bpf() if user is not privileged

On Linux, bpf() returns EPERM if the user doesn't have CAP_BPF. Return
the same error when the user isn't able to open the device handle.