Return ENOENT from array maps containing object references (#3933)
A lookup in an array map yields all zeroes if the index of the array is
empty. This makes sense in the general case, but has surprising semantics
when applied to nested maps like arrays of arrays.
In this case, looking up an empty slot will succeed but the all zero value
is interpreted as a zero ID, which causes an error when trying to create
a map from that ID.

Follow the Linux semantics which return ENOENT in this special case.