add .vsconfig
Store dependencies in a .vsconfig file, which is picked up by Visual
Studio when opening the solution file. This is easier than manually
selecting the correct options in the installer.

Change Setup-DevEnv.ps1 to use the .vsconfig file from the repo and
remove the buildtools installation, since those seem to be duplicates.