bpf(): improve map and program compatibility (#3980)
* bpf(): improve map and program compatibility

Allow passing a name for maps and programs and allow retrieving it via
BPF_OBJ_GET_INFO_BY_FD. This requires a bunch of glue code because the
bpf_prog_info, etc. structs are not compatible with Linux.

* Fix test failure due to incorrect link type

* Address feedback from Dave Thaler

* Fix second test failure due to wrong link type

* Use program ID 0 on detached link ids

Internally both zero and EBPF_ID_NONE are used to refer to non-existing ID.
Use zero when retrieving link info, since that aligns with Linux better.

* Temporarily use BPF_LINK_TYPE_UNSPEC

* Only test program info if JIT is enabled

* Apply suggestions from dthaler's review

Co-authored-by: Dave Thaler <dthaler1968@gmail.com>

---------

Co-authored-by: Dave Thaler <dthaler1968@gmail.com>