Add ebpf_get_next_pinned_object_path
It's currently not possible to find pinned maps and links. Add a
function ebpf_get_next_pinned_object_path which allows doing just
that.

The paths are returned in lexicographical order, which allows user
space to performa a prefix match on the paths. Behind the scenes
this is implemented using a linear scan of the hash table, which
has quadratic running time. The same is already the case for finding
the next object ID, so going with the more desirable semantics
makes sense.

The existing ebpf_get_next_pinned_program_path() is deprecated because
it is less flexible than the new function.

Updates https://github.com/microsoft/ebpf-for-windows/issues/4161