Remove EBPF_ENABLE_WER_REPORT in favour of IsDebuggerPresent()
The unit tests inspect EBPF_ENABLE_WER_REPORT environment variable to
determine whether to customise error reporting. By default, no
customisation is done.
On headless environments like CI / ssh sessions this leads to the program
hanging or simply being killed.

Replace the environment variable with a call to IsDebuggerPresent():
if there is a debugger we bail out. Otherwise we configure error
reporting to print assertions to stderr before crashing.

This allows removing a bunch of boilerplate from CI and reduces the
amount of headscratching when a test fails without any output.