bpf: support log_true_size for BPF_PROG_LOAD (#4360)
* Add log_buffer_true_size argument to ebpf_program_load_bytes

* Fix off-by-one in calculation of error_message_size in ebpf_verify_and_load_program

* bpf: support log_true_size for BPF_PROG_LOAD

The behaviour around verifier logs on Linux is quite subtle and a bit
error prone. One API that makes everything a lot easier is the
log_true_size field which informs the caller how big the provided
buffer should be. Implement support for the field in the bpf()
syscall wrapper.